+++ CHANGES.md
@@ -18,6 +18,7 @@ Changes in 4.7
 - Automatic Class Rank calculation in EditReportCardGrades.php, GPARankList.php, InputFinalGrades.php, Transcripts.php
 - SQL performance: rewrite set_class_rank_mp() function in rosariosis.sql & Update.fnc.php
 - SQL remove calc_cum_gpa_mp() function & include it in t_update_mp_stats() trigger in rosariosis.sql & Update.fnc.php
+- Remove Calculate GPA program in CalcGPA.php, Menu.php, Help_en.php & rosariosis.sql
 
 Changes in 4.6.2
 ----------------